+++ src/main/java/school/sptech/backend/service/tipocesta/dto/TipoCestaListagemDto.java
@@ -7,13 +7,13 @@ import java.util.List;
 
 @Data
 public class TipoCestaListagemDto {
-    private Long id;
+    private Integer id;
     private String nome;
     private List<CestaDto> cestas;
 
     @Data
     public static class CestaDto{
-        private Long id;
+        private Integer id;
         private String lote;
         private LocalDate dataMontagem;
     }
